refactor: this change comments pagination logic
